Update route.ts
Get ,Add and Delete task for user is working as expected

+++ todo-next-drizzle-sanity/app/api/todo/route.ts
@@ -1,7 +1,7 @@
 import { NextRequest, NextResponse } from "next/server";
 import { db, todoTable } from "@/app/lib/drizzle";
 // import { sql } from "@vercel/postgres";
-import { eq } from "drizzle-orm";
+import { and, eq } from "drizzle-orm";
 import { cookies } from "next/headers";
 import { v4 as uuid } from "uuid";
 
@@ -40,9 +40,10 @@ export const POST = async (request: NextRequest) => {
 };
 
 export async function DELETE(request: NextRequest) {
+  const user_id=cookies().get("user_id")?.value as string
   try {
     const id: any = request.nextUrl.searchParams.get("id");
-    await db.delete(todoTable).where(eq(todoTable.id, id)).execute();
+    await db.delete(todoTable).where(and(eq(todoTable.id, id),eq(todoTable.user_id,user_id))).execute();
     return NextResponse.json({ message: "Todo deleted successfully" });
   } catch (error) {
     console.log("error :>> ", error);
